extend controller example; add form error example

+++ src/main/resources/archetype-resources/src/main/java/controller/AppController.java
@@ -9,6 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -19,7 +20,7 @@ import ${package}.domain.AppRepository;
 @Controller
 @RequestMapping("/apps")
 public class AppController {
-	
+
 	private static final transient Logger LOG = LoggerFactory.getLogger(AppController.class);
 	@Inject
 	private AppRepository appRepository;
@@ -32,9 +33,12 @@ public class AppController {
 	}
 
 	@RequestMapping(method = RequestMethod.GET, value = "/{id}")
-	public String show(@PathVariable final Long id, final Model model) {
+	public String show(@PathVariable final Integer id, final Model model) {
 		LOG.debug("show");
 		App app = appRepository.findById(id);
+		if (app == null) {
+			return "redirect:/404.html";
+		}
 		model.addAttribute("app", app);
 		return "app/show";
 	}
@@ -47,34 +51,57 @@ public class AppController {
 		model.addAttribute("app", app);
 		return "app/new";
 	}
-	
+
 	@RequestMapping(method = RequestMethod.GET, value = "/{id}/edit")
-	public String edit(@PathVariable final Long id, final Model model) {
+	public String edit(@PathVariable final Integer id, final Model model) {
 		LOG.debug("edit");
 		App app = appRepository.findById(id);
+		if (app == null) {
+			return "redirect:/404.html";
+		}
 		model.addAttribute("app", app);
 		return "app/edit";
 	}
 
 	@RequestMapping(method = RequestMethod.POST)
-	public String save(final App app) {
+	public String save(final App newApp, final BindingResult result) {
 		LOG.debug("save");
-		appRepository.save(app);
+		if (appRepository.findByName(newApp.getName()) != null) {
+			result.rejectValue("name", "app.error.alreadyexits.name");
+		}
+		if (result.hasErrors()) {
+			return "app/new";
+		}
+		appRepository.save(newApp);
 		return "redirect:/apps";
 	}
 
 	@RequestMapping(method = RequestMethod.PUT, value = "/{id}")
-	public String update(@PathVariable final Long id, final App updatedApp) {
+	public String update(@PathVariable final Integer id, final App updatedApp, final BindingResult result) {
 		LOG.debug("update");
-		App app = appRepository.findById(id);
-		app.setName(updatedApp.getName());
-		appRepository.save(app);
+		App existingApp = appRepository.findById(id);
+		if (existingApp == null) {
+			return "redirect:/404.html";
+		}
+		App byName = appRepository.findByName(updatedApp.getName());
+		if ((byName != null) && !byName.getId().equals(id)) {
+			result.rejectValue("name", "app.error.alreadyexits.name");
+		}
+		if (result.hasErrors()) {
+			return "app/edit";
+		}
+		existingApp.setName(updatedApp.getName());
+		appRepository.save(existingApp);
 		return "redirect:/apps";
 	}
-	
+
 	@RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
-	public String delete(@PathVariable final Long id) {
+	public String delete(@PathVariable final Integer id) {
 		LOG.debug("delete");
+		App existingApp = appRepository.findById(id);
+		if (existingApp == null) {
+			return "redirect:/404.html";
+		}
 		appRepository.remove(id);
 		return "redirect:/apps";
 	}
